Jwt_api/serializer.py
from django.contrib.auth.models import User, Group
from rest_framework import serializers


class UserSerializer(serializers.HyperlinkedModelSerializer):

    class Meta:
        model = User
        fields = ( 'username', 'email', 'password')

    def create(self, validated_data):
        # override create function to save password
        user = User.objects.create(username=validated_data['username'],email=validated_data['email'] )
        user.set_password(validated_data['password'])
        user.save()
        return user


class GroupSerializer(serializers.HyperlinkedModelSerializer):
    class Meta:
        model = Group
        fields = ('url', 'name')