requirements.txt
defusedxml==0.4.1
Django==1.9.7
django-allauth==0.25.2
django-social-auth==0.7.28
django-twitter-bootstrap==3.3.0
djangorestframework==3.3.3
djangorestframework-jwt==1.8.0
httplib2==0.9.2
oauth2==1.9.0.post1
oauthlib==1.1.2
PyJWT==1.4.0
python-openid==2.2.5
python3-openid==3.0.10
requests==2.10.0
requests-oauthlib==0.6.1
